+++ src/main/java/com/example/demo/DTOs/GioHangDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.*;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -14,11 +16,11 @@ public class GioHangDTO {
 
     @NotNull(message = "Số tiền giảm không được để trống")
     @DecimalMin(value = "0.0", message = "Số tiền giảm phải lớn hơn hoặc bằng 0")
-     Double soTienGiam;
+    BigDecimal soTienGiam;
 
     @NotNull(message = "Tổng tiền không được để trống")
     @DecimalMin(value = "0.0", message = "Tổng tiền phải lớn hơn hoặc bằng 0")
-     Double tongTien;
+     BigDecimal tongTien;
 
     @NotBlank(message = "Trạng thái không được để trống")
     @Size(max = 20, message = "Trạng thái không được vượt quá 20 ký tự")
